# api/lifecycle.openapi.yaml
openapi: 3.1.0
info:
  title: MiniSandbox Lifecycle API
  version: 0.2.0
servers:
  - url: http://127.0.0.1:8080
paths:
  /healthz:
    get:
      operationId: health
      responses:
        "200":
          description: Control plane is healthy
  /readyz:
    get:
      operationId: ready
      responses:
        "200":
          description: Control plane is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
        "503":
          description: One or more required control-plane components are not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
  /v1/sandboxes:
    post:
      operationId: createSandbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSandboxRequest"
      responses:
        "202":
          description: Sandbox creation was accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/sandboxes/{sandbox_id}:
    parameters:
      - $ref: "#/components/parameters/SandboxID"
    get:
      operationId: getSandbox
      responses:
        "200":
          description: Current sandbox state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sandbox"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      operationId: deleteSandbox
      responses:
        "202":
          description: Sandbox deletion was accepted
        "204":
          description: Sandbox is already absent
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/sandboxes/{sandbox_id}/executions:
    parameters:
      - $ref: "#/components/parameters/SandboxID"
    post:
      operationId: executeSandboxCommand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteRequest"
            examples:
              foregroundArgv:
                value:
                  argv: [go, test, ./...]
                  cwd: /workspace
                  env:
                    CI: "true"
                  timeout_seconds: 120
                  background: false
      responses:
        "200":
          description: Foreground execution event stream ending in exactly one terminal event
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE frames whose data JSON conforms to ExecutionEvent
        "202":
          description: Background execution was accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionDescriptor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/sandboxes/{sandbox_id}/executions/{execution_id}:
    parameters:
      - $ref: "#/components/parameters/SandboxID"
      - $ref: "#/components/parameters/ExecutionID"
    get:
      operationId: getSandboxExecution
      responses:
        "200":
          description: Current execution state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      operationId: cancelSandboxExecution
      responses:
        "202":
          description: Cancellation of a running execution was accepted
        "204":
          description: Execution is already terminal
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/sandboxes/{sandbox_id}/executions/{execution_id}/logs:
    parameters:
      - $ref: "#/components/parameters/SandboxID"
      - $ref: "#/components/parameters/ExecutionID"
    get:
      operationId: getSandboxExecutionLogs
      parameters:
        - $ref: "#/components/parameters/LogCursor"
      responses:
        "200":
          description: Execution events after the supplied sequence cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionLogPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  parameters:
    SandboxID:
      name: sandbox_id
      in: path
      required: true
      schema:
        type: string
    ExecutionID:
      name: execution_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    LogCursor:
      name: cursor
      in: query
      required: false
      description: Last event sequence already read; omitted is equivalent to zero
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Request is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Sandbox does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Server policy forbids the requested capability
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current sandbox state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An unexpected internal error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: A required control-plane dependency is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Execution cannot be started with the supplied semantic values
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Execution concurrency limit has been reached
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      example:
        error:
          code: INVALID_REQUEST
          message: Request is invalid.
          request_id: req-01
          retryable: false
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
          description: Stable machine-readable error code
        message:
          type: string
          description: Safe error detail without secrets or internal causes
        request_id:
          type: string
          description: Request identifier also returned in the X-Request-ID header
        retryable:
          type: boolean
          description: Whether retrying later may succeed without changing the request
    ErrorCode:
      type: string
      enum:
        - INVALID_REQUEST
        - SANDBOX_NOT_FOUND
        - SANDBOX_CONFLICT
        - SANDBOX_NOT_RUNNING
        - INVALID_EXECUTION_REQUEST
        - EXECUTION_NOT_FOUND
        - EXECUTION_LIMIT_REACHED
        - SHELL_NOT_FOUND
        - INVALID_CWD
        - RUNNER_UNHEALTHY
        - RUNNER_PROTOCOL_MISMATCH
        - OUTBOUND_NOT_ALLOWED
        - EGRESS_IMAGE_UNAVAILABLE
        - EGRESS_POLICY_INVALID
        - EGRESS_NOT_READY
        - EGRESS_UNHEALTHY
        - RUNTIME_UNAVAILABLE
        - NOT_IMPLEMENTED
        - INTERNAL_ERROR
    CreateSandboxRequest:
      type: object
      additionalProperties: false
      required: [image]
      properties:
        image:
          type: string
        network:
          $ref: "#/components/schemas/SandboxNetworkRequest"
    SandboxNetworkRequest:
      type: object
      additionalProperties: false
      properties:
        outbound:
          type: boolean
          default: false
    ExecuteRequest:
      type: object
      additionalProperties: false
      properties:
        argv:
          type: array
          minItems: 1
          prefixItems:
            - type: string
              minLength: 1
          items:
            type: string
        shell:
          type: string
          minLength: 1
        cwd:
          type: string
          description: Container path below /workspace; defaults to /workspace
        env:
          type: object
          additionalProperties:
            type: string
        timeout_seconds:
          type: integer
          format: int64
          minimum: 0
          description: Execution timeout in seconds; zero selects the runner default
        background:
          type: boolean
          default: false
      oneOf:
        - required: [argv]
        - required: [shell]
    ExecutionState:
      type: string
      enum: [Pending, Running, Exited, Failed, Cancelled, TimedOut]
    ExecutionDescriptor:
      type: object
      additionalProperties: false
      required: [execution_id, state]
      properties:
        execution_id:
          type: string
        state:
          $ref: "#/components/schemas/ExecutionState"
    ExecutionStatus:
      type: object
      additionalProperties: false
      required: [execution_id, state]
      properties:
        execution_id:
          type: string
        state:
          $ref: "#/components/schemas/ExecutionState"
        terminal_event:
          $ref: "#/components/schemas/TerminalExecutionEvent"
    ExecutionLogPage:
      type: object
      additionalProperties: false
      required: [events, next_cursor, complete]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/ExecutionEvent"
        next_cursor:
          type: integer
          format: int64
          minimum: 0
        complete:
          type: boolean
    ExecutionEvent:
      oneOf:
        - $ref: "#/components/schemas/StartedEvent"
        - $ref: "#/components/schemas/StdoutEvent"
        - $ref: "#/components/schemas/StderrEvent"
        - $ref: "#/components/schemas/OutputLimitReachedEvent"
        - $ref: "#/components/schemas/ExitedEvent"
        - $ref: "#/components/schemas/FailedEvent"
        - $ref: "#/components/schemas/CancelledEvent"
        - $ref: "#/components/schemas/TimedOutEvent"
      discriminator:
        propertyName: type
    TerminalExecutionEvent:
      oneOf:
        - $ref: "#/components/schemas/ExitedEvent"
        - $ref: "#/components/schemas/FailedEvent"
        - $ref: "#/components/schemas/CancelledEvent"
        - $ref: "#/components/schemas/TimedOutEvent"
      discriminator:
        propertyName: type
    StartedEvent:
      type: object
      additionalProperties: false
      required: [execution_id, sequence, timestamp, type]
      properties:
        execution_id: {type: string}
        sequence: {type: integer, format: int64, minimum: 1}
        timestamp: {type: string, format: date-time}
        type: {type: string, const: started}
    OutputLimitReachedEvent:
      type: object
      additionalProperties: false
      required: [execution_id, sequence, timestamp, type]
      properties:
        execution_id: {type: string}
        sequence: {type: integer, format: int64, minimum: 1}
        timestamp: {type: string, format: date-time}
        type: {type: string, const: output_limit_reached}
    StdoutEvent:
      type: object
      additionalProperties: false
      required: [execution_id, sequence, timestamp, type, data_base64]
      properties:
        execution_id: {type: string}
        sequence: {type: integer, format: int64, minimum: 1}
        timestamp: {type: string, format: date-time}
        type: {type: string, const: stdout}
        data_base64: {type: string, format: byte}
    StderrEvent:
      type: object
      additionalProperties: false
      required: [execution_id, sequence, timestamp, type, data_base64]
      properties:
        execution_id: {type: string}
        sequence: {type: integer, format: int64, minimum: 1}
        timestamp: {type: string, format: date-time}
        type: {type: string, const: stderr}
        data_base64: {type: string, format: byte}
    ExitedEvent:
      type: object
      additionalProperties: false
      required: [execution_id, sequence, timestamp, type, exit_code, duration_ms, output_truncated]
      properties:
        execution_id: {type: string}
        sequence: {type: integer, format: int64, minimum: 1}
        timestamp: {type: string, format: date-time}
        type: {type: string, const: exited}
        exit_code: {type: integer}
        duration_ms: {type: integer, format: int64, minimum: 0}
        output_truncated: {type: boolean}
    FailedEvent:
      type: object
      additionalProperties: false
      required: [execution_id, sequence, timestamp, type, error_code, message, duration_ms, output_truncated]
      properties:
        execution_id: {type: string}
        sequence: {type: integer, format: int64, minimum: 1}
        timestamp: {type: string, format: date-time}
        type: {type: string, const: failed}
        error_code: {type: string}
        message: {type: string}
        duration_ms: {type: integer, format: int64, minimum: 0}
        output_truncated: {type: boolean}
    CancelledEvent:
      type: object
      additionalProperties: false
      required: [execution_id, sequence, timestamp, type, duration_ms, output_truncated]
      properties:
        execution_id: {type: string}
        sequence: {type: integer, format: int64, minimum: 1}
        timestamp: {type: string, format: date-time}
        type: {type: string, const: cancelled}
        duration_ms: {type: integer, format: int64, minimum: 0}
        output_truncated: {type: boolean}
    TimedOutEvent:
      type: object
      additionalProperties: false
      required: [execution_id, sequence, timestamp, type, duration_ms, output_truncated]
      properties:
        execution_id: {type: string}
        sequence: {type: integer, format: int64, minimum: 1}
        timestamp: {type: string, format: date-time}
        type: {type: string, const: timed_out}
        duration_ms: {type: integer, format: int64, minimum: 0}
        output_truncated: {type: boolean}
    ReadinessStatus:
      type: string
      enum: [ready, not_ready]
    ReadinessComponentName:
      type: string
      enum: [store, docker, artifact, recovery, worker]
    ReadinessComponent:
      type: object
      additionalProperties: false
      required: [name, status]
      properties:
        name:
          $ref: "#/components/schemas/ReadinessComponentName"
        status:
          $ref: "#/components/schemas/ReadinessStatus"
    ReadinessResponse:
      type: object
      additionalProperties: false
      required: [status, components]
      properties:
        status:
          $ref: "#/components/schemas/ReadinessStatus"
        components:
          type: array
          minItems: 5
          maxItems: 5
          items:
            $ref: "#/components/schemas/ReadinessComponent"
    SandboxState:
      type: string
      enum:
        - Pending
        - Creating
        - Running
        - Stopping
        - Terminated
        - Failed
    SandboxReason:
      type: string
      enum:
        - CREATE_ACCEPTED
        - CREATING_RUNTIME
        - WAITING_RUNNER
        - RUNNING
        - DELETE_ACCEPTED
        - DELETING_RUNTIME
        - TERMINATED
        - IMAGE_PULL_FAILED
        - ARTIFACT_INVALID
        - CONTAINER_CREATE_FAILED
        - ARTIFACT_INJECTION_FAILED
        - CONTAINER_START_FAILED
        - RUNNER_UNHEALTHY
        - SPEC_DRIFT
        - CLEANUP_PENDING
        - RUNTIME_UNAVAILABLE
        - INTERNAL_ERROR
    Sandbox:
      type: object
      additionalProperties: false
      required: [id, state, reason, message, image, created_at, updated_at]
      properties:
        id:
          type: string
        state:
          $ref: "#/components/schemas/SandboxState"
        reason:
          $ref: "#/components/schemas/SandboxReason"
        message:
          type: string
          description: Safe lifecycle detail without secrets or host paths
        image:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: sbx-01
        state: Pending
        reason: CREATE_ACCEPTED
        message: Sandbox creation has been accepted.
        image: alpine:3.22
        created_at: "2026-07-25T10:30:00Z"
        updated_at: "2026-07-25T10:30:00Z"